+++ src/components/Users/Personal-data/Personal-data.tsx
@@ -5,9 +5,10 @@ import { GET_ME } from "@/Graphql/Schemas/UserQuery";
 import { useAppDispatch } from '@/libs/Store/Store';
 import { deleteUser } from '@/Features/Actions/usersActions';
 import { useState } from 'react';
-import EditUserDataForm from '@/utils/Forms/Edit-User-Data-Form';
+import EditUserDataForm from '@/utils/Forms/User/Edit-User-Data-Form';
 import Image from 'next/image';
-import ChangePassword from '../../../utils/Forms/ChangePassword';
+import ChangePassword from '../../../utils/Forms/User/ChangePassword';
+import swal from 'sweetalert';
 
 
 export default function Personal_Data() {
@@ -20,8 +21,27 @@ export default function Personal_Data() {
   const dispatch = useAppDispatch()
   //Handller Delete User
   const DeleteUserHandller = (id: string) => {
-    dispatch(deleteUser(id))
-    window.location.href = '/';
+
+    swal({
+            title: "هل أنت متأكد من الحذف؟",
+            text: "بمجرد حذف التصنيف سيتم مسحة نهائياً ولن تستطع إستعادة بياناته مرة أخرى!",
+            icon: "warning",
+            dangerMode: true,
+            buttons:["إلغاء", "حذف"]
+          })
+        .then((willDelete) => {
+          if (willDelete) {
+            dispatch(deleteUser(id))
+            swal("تم حذف التصنيف", {
+              icon: "success",
+            });
+            window.location.href = '/';
+          } else {
+            swal("أنت الأن في أمان لم يتم الحذف!");
+          }
+        });
+    
+    
   }
   return (
     <div className="max-w-5xl mx-auto text-foreground w-full border relative p-6 rounded-lg shadow">
@@ -60,19 +80,19 @@ export default function Personal_Data() {
           </div >
           <div className='flex flex-wrap gap-2'>
             <label className="text-sm text-gray-500">تاريخ الميلاد</label>
-            <p className="font-medium">{data?.me?.birthDate?(new Date(Number(data?.me?.birthDate)).toLocaleString('ar-EG', {
+            <p className="font-medium">{data?.me?.birthDate ? (new Date(Number(data?.me?.birthDate)).toLocaleString('ar-EG', {
               year: 'numeric',
               month: 'short',
               day: '2-digit'
-            })):'جارٍ التحميل .....'
-            
+            })) : 'جارٍ التحميل .....'
+
             }</p>
           </div>
         </div>
         {/*Edit User Data Form*/}
         {
-          isEdit 
-            && (
+          isEdit
+          && (
             <EditUserDataForm setIsEdit={setIsEdit} />
           )
         }
@@ -80,7 +100,7 @@ export default function Personal_Data() {
         {
           isChangePassword && (
             <ChangePassword setIsChangePassword={setIsChangePassword} />
-            )
+          )
         }
         <div className="flex gap-1 mt-6 w-full ">
           <button onClick={() => setIsEdit(!isEdit)} className="bg-blue-600 text-white px-4 py-2 cursor-pointer rounded hover:bg-blue-700">
@@ -89,7 +109,9 @@ export default function Personal_Data() {
           <button onClick={() => setIsChangePassword(!isChangePassword)} className="bg-gray-100 text-gray-800 px-4 py-2 cursor-pointer rounded hover:bg-gray-200">
             تغيير كلمة المرور
           </button>
-          <button onClick={() => DeleteUserHandller(data?.me?.id)} className='text-white hover:bg-red-600 rounded  mr-auto  bg-sky-500 p-2  cursor-pointer'>حذف الحساب</button>
+          {data?.me?.role !== 'ADMIN' && <button onClick={() => DeleteUserHandller(data?.me?.id)} className='text-white hover:bg-red-600 rounded  mr-auto  bg-sky-500 p-2  cursor-pointer'>
+            حذف الحساب
+            </button>}
         </div>
       </div>
     </div>
